# server/docs/swagger.yaml
...
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for Schej.it!
  title: Schej.it API
  version: "1.0"
paths:
...
  /polar/billing-portal:
    get:
      parameters:
      - description: URL to return to from the portal
        in: query
        name: returnUrl
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              url:
                type: string
            type: object
      summary: Gets a Polar customer portal URL for the signed in user
      tags:
      - polar
  /polar/create-checkout-session:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product to purchase and URL to return to afterwards
        in: body
        name: payload
        required: true
        schema:
          properties:
            originUrl:
              type: string
            productId:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              url:
                type: string
            type: object
      summary: Creates a Polar checkout session for the signed in user
      tags:
      - polar
  /polar/fulfill-checkout:
    post:
      consumes:
      - application/json
      description: Returns 202 if the payment is still being processed; the client
        should retry.
      parameters:
      - description: Polar checkout ID
        in: body
        name: payload
        required: true
        schema:
          properties:
            checkoutId:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: status is \"fulfilled\
          schema:
            properties:
              status:
                type: string
            type: object
        "202":
          description: status is \"pending\
          schema:
            properties:
              status:
                type: string
            type: object
      summary: Fulfills a completed Polar checkout
      tags:
      - polar
  /polar/price:
    get:
      parameters:
      - description: Pricing experiment variant
        in: query
        name: exp
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              monthly:
                type: object
              monthlyStudent:
                type: object
              yearly:
                type: object
              yearlyStudent:
                type: object
            type: object
      summary: Gets the prices of all premium plans
      tags:
      - polar
...
swagger: "2.0"